LPS-114698 Rename get it close to some of: git grep "@Produces(MediaType.APPLICATION_JSON)" "**.java"But I couldn't identify a patterns for files like this that are not generated so I'm putting this into /jaxrs/endpoint but /jaxrs/resource or just /resource could also makes sense
